[DF-7045] build fix
Former-commit-id: 8c40ad29708ee8c12a5ffdf2a625e74762cd0d1e

+++ Sources/Internal/Platform/TemplateAndroid/DeviceInfoAndroid.cpp
@@ -124,7 +124,7 @@ WideString DeviceInfoPrivate::GetName()
 
 int32 DeviceInfoPrivate::GetZBufferSize()
 {
-	return getZBufferSize();
+	return static_cast<int32>(getZBufferSize());
 }
 
 String DeviceInfoPrivate::GetHTTPProxyHost()
@@ -139,7 +139,7 @@ String DeviceInfoPrivate::GetHTTPNonProxyHosts()
 
 int32 DeviceInfoPrivate::GetHTTPProxyPort()
 {
-	return getHTTPProxyPort();
+	return static_cast<int32>(getHTTPProxyPort());
 }
 
 DeviceInfo::ScreenInfo& DeviceInfoPrivate::GetScreenInfo()
